main.py
import logging
import os
import sqlite3
from datetime import datetime
import re
import json
from collections import Counter
import emoji
import numpy as np
from telegram import Update, ReplyKeyboardMarkup, KeyboardButton, ReplyKeyboardRemove
from telegram.ext import Application, CommandHandler, MessageHandler, filters, ContextTypes, ConversationHandler
import requests
from dotenv import load_dotenv
import nltk
from nltk.corpus import stopwords
from nltk.tokenize import word_tokenize
from nltk.stem import SnowballStemmer
from textblob import TextBlob
from langdetect import detect
from sentence_transformers import SentenceTransformer
import threading
from flask import Flask

# Загружаем переменные окружения
load_dotenv()

# ... (остальной код из вашего запроса остается без изменений до функции main)

def keep_alive():
    """Функция для постоянной работы в Railway.app"""
    app = Flask(__name__)
    
    @app.route('/')
    def home():
        return "✅ Бот активен и работает!"
    
    def run():
        app.run(host='0.0.0.0', port=int(os.environ.get('PORT', 8080)))
    
    threading.Thread(target=run, daemon=True).start()
    logger.info(f"✅ HTTP сервер для keep-alive запущен на порту {os.environ.get('PORT', 8080)}")

def main():
    """Основная функция запуска бота"""
    # Запускаем keep-alive сервер для Railway
    keep_alive()
    
    logger.info("🚀 Запуск Telegram бота с интеллектуальным поиском...")
    
    # Инициализация базы данных
    init_database()
    
    # Загрузка NLP моделей
    preload_nlp_models()
    
    BOT_TOKEN = os.getenv('BOT_TOKEN')
    if not BOT_TOKEN:
        logger.critical("❌ BOT_TOKEN не найден в переменных окружения!")
        logger.critical("Пожалуйста, установите BOT_TOKEN в Railway Variables")
        return
    
    try:
        application = Application.builder().token(BOT_TOKEN).build()
        
        # Добавляем обработчики
        conv_handler = ConversationHandler(
            # ... (остальной код без изменений)
        )
        
        application.add_handler(conv_handler)
        application.add_handler(CommandHandler('help', help_command))
        application.add_handler(CommandHandler('groups', groups_command))
        application.add_handler(CommandHandler('support', support_command))
        
        logger.info("✅ Бот успешно инициализирован")
        logger.info("⚡ Бот запущен и готов к приему сообщений!")
        
        # Запускаем в режиме polling
        application.run_polling(drop_pending_updates=True)
        
    except Exception as e:
        logger.critical(f"🔥 КРИТИЧЕСКАЯ ОШИБКА: {e}")
        logger.critical("Проверьте правильность настроек в Railway")

if __name__ == "__main__":
    main()