requirements.txt
python-telegram-bot==20.7
python-dotenv==1.0.0
flask==3.0.0
requests==2.31.0
nltk==3.8.1
numpy==1.24.3
textblob==0.17.1
emoji==2.8.0
langdetect==1.0.9
sentence-transformers==2.2.2
torch==2.0.1
transformers==4.30.2
python-telegram-bot==20.7
python-dotenv==1.0.0
flask==3.0.0
requests==2.31.0
nltk==3.8.1
numpy==1.24.3
textblob==0.17.1
emoji==2.8.0
langdetect==1.0.9
sentence-transformers==2.2.2
torch==2.0.1
transformers==4.30.2